==> Src/CNN-RNN/01_Video_feature_extraction.ipynb <==
## test

import tensorflow as tf
import os

os.environ["CUDA_VISIBLE_DEVICES"] = "0"
print(tf.__version__)
print(tf.config.list_physical_devices('GPU'))
#tf.config.experimental_run_functions_eagerly(True)

## import

import pandas as pd
import numpy as np
from tensorflow.keras.preprocessing import image
from tensorflow.keras.utils import plot_model

## feature extractor

network_name = 'ResNet50 lts not preprocessed'

def build_feature_extractor():

    feature_extractor = tf.keras.applications.ResNet50(weights = "resnet50_weights_tf_dim_ordering_tf_kernels_notop.h5",
                                                       include_top = False, pooling = "avg", input_shape = (
        400, 300, 3))
    #preprocess_input  = tf.keras.applications.inception_v3.preprocess_input
    inputs = tf.keras.Input((400, 300, 3))
    #preprocessed      = preprocess_input(inputs)
    outputs = feature_extractor(inputs)

    return tf.keras.Model(inputs, outputs, name = "feature_extractor")


feature_extractor = build_feature_extractor()

n = os.listdir('I:/poorya/datasets/video dataset/videos lts/')
features = np.zeros((len(n), 40, 2048), dtype = "float32")
strsr_lbl = np.zeros((len(n), 2))
typee_lbl = np.zeros((len(n), 4))

for i in range(len(n)):
    strsr_lbl[i, int(n[i][-3])] = 1
    typee_lbl[i, int(n[i][-1])] = 1
    print(i)

    for j in range(40):
        img = image.img_to_array(image.load_img('I:/poorya/datasets/video dataset/videos lts/' + n[
            i] + '/' + str(j) + '.jpg'))
        img /= 255.0
        features[i, j, :] = feature_extractor.predict(np.expand_dims(img, 0)).squeeze()

#strsr_lbl = np.load('I:/poorya/datasets/video dataset/lables/strsr_lbl.npy'             ) 
#ll = np.zeros((len(strsr_lbl)))

#for i in range(len(strsr_lbl)) :
#ll[i] = int(np.where(np.array(strsr_lbl[i]) == 1)[0][0])

#ll = ll.astype('int')

os.mkdir('I:/poorya/datasets/video dataset/' + network_name)
np.save('I:/poorya/datasets/video dataset/' + network_name + '/features.npy', features)
np.save('I:/poorya/datasets/video dataset/lables/strsr_lts_lbl.npy', strsr_lbl)
np.save('I:/poorya/datasets/video dataset/lables/typee_lts_lbl.npy', typee_lbl)
#np.save( 'I:/poorya/datasets/video dataset/lables/ll.npy'                     , ll       )

## data preparation

#features  = np.load('I:/poorya/datasets/video dataset/' + network_name + '/features.npy')
#typee_lbl = np.load('I:/poorya/datasets/video dataset/lables/typee_lbl.npy'             )
#ll        = np.load('I:/poorya/datasets/video dataset/lables/ll.npy'                    )

l = len(features)
values = [i for i in range(l)]
permutations = np.random.permutation(values)
X = features[permutations, :, :]
Ys = strsr_lbl[permutations]

X_train = X[:int(0.8 * l), :, :]
Y_train = Ys[:int(0.8 * l)]
X_test = X[int(0.8 * l):, :, :]
Y_test = Ys[int(0.8 * l):]

sizee = len(X_test)
val = 0.5
X_val = X_test[:int(val * sizee), :]
Y_val = Y_test[:int(val * sizee)]
X_test = X_test[int(val * sizee):, :]
Y_test = Y_test[int(val * sizee):]

tr_df = pd.DataFrame({
                         'file name':      [str(i) + '.npy' for i in range(len(X_train))],
                         'stressor lable': [[int(Y_train[i, 0]), int(Y_train[i, 1])] for i in range(len(Y_train))]
                         })

ts_df = pd.DataFrame({
                         'file name':      [str(i) + '.npy' for i in range(len(X_test))],
                         'stressor lable': [[int(Y_test[i, 0]), int(Y_test[i, 1])] for i in range(len(Y_test))]
                         })

vl_df = pd.DataFrame({
                         'file name':      [str(i) + '.npy' for i in range(len(X_val))],
                         'stressor lable': [[int(Y_val[i, 0]), int(Y_val[i, 1])] for i in range(len(Y_val))]
                         })

#'stressor lable': [[int(Y_train[i,0]), int(Y_train[i,1])] for i in range(len(Y_train))]})

os.mkdir('I:/poorya/datasets/video dataset/' + network_name + '/train')
os.mkdir('I:/poorya/datasets/video dataset/' + network_name + '/test')
os.mkdir('I:/poorya/datasets/video dataset/' + network_name + '/val')

tr_df.to_csv('I:/poorya/datasets/video dataset/' + network_name + '/train/tr_df.csv')
ts_df.to_csv('I:/poorya/datasets/video dataset/' + network_name + '/test/ts_df.csv')
vl_df.to_csv('I:/poorya/datasets/video dataset/' + network_name + '/val/vl_df.csv')

def sv(myx, adrs):
    for i in range(len(myx)):
        np.save('I:/poorya/datasets/video dataset/' + network_name + '/' + adrs + str(i) + '.npy', myx[i, :, :])


sv(X_train, 'train/')
sv(X_test, 'test/')
sv(X_val, 'val/')
